closing-steps-orchestration

+++ src/Modules/Closing/Domain/TrustYields/ITrustYieldRepository.cs
@@ -2,7 +2,7 @@ namespace Closing.Domain.TrustYields;
 
 public interface ITrustYieldRepository
 {
-    Task<TrustYield?> GetByTrustAndDateAsync(int trustId, DateTime closingDateUtc, CancellationToken cancellationToken = default);
+    Task<TrustYield?> GetByTrustAndDateAsync(long trustId, DateTime closingDateUtc, CancellationToken cancellationToken = default);
     Task InsertAsync(TrustYield trustYield, CancellationToken cancellationToken = default);
     void Update(TrustYield trustYield);
     Task<IReadOnlyCollection<TrustYield>> GetByPortfolioAndDateAsync(int portfolioId, DateTime closingDate, CancellationToken ct);
